Changed the color of copy button
The primary button color is a light color and the on copy color is a
blue color now. This should really help when the button is clicked and
switches to the on copy color. Since the button is located in the card's
header and the card can be either red or green, the on copy color would
have blended into the card header.